[2020-06-18] util 패키지 코드 및 구조 수정
