--- app/src/main/java/com/memo/project/util/OnItemSelectedListener.kt
@@ -1,6 +1,0 @@
-package com.memo.project.util
-
-import android.view.View
-
-typealias OnItemSelectedListener<T> = ((view : View, item : T, position : Int) -> Unit)
-typealias OnImageSelectedListener = ((view : View, imagePath : String) -> Unit)
