+++ app/src/main/java/com/memo/project/util/SelectedListener.kt
@@ -1,0 +1,5 @@
+package com.memo.project.util
+
+typealias OnItemSelectedListener = ((index: Long) -> Unit)
+
+typealias OnImageSelectedListener = ((imagePath : String) -> Unit)
