+++ app/src/main/java/com/memo/project/util/extenstion/ViewModel.kt
@@ -3,6 +3,6 @@ package com.memo.project.util.extenstion
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-inline fun <reified T> ViewModel.createMutableLiveData(defaultValue: T? = null): Lazy<MutableLiveData<T>> = lazy{
-    MutableLiveData<T>(defaultValue)
-}
+inline fun <reified T> ViewModel.createMutableLiveData(
+    defaultValue: T? = null
+): Lazy<MutableLiveData<T>> = lazy{ MutableLiveData<T>(defaultValue) }
