+++ app/src/main/java/com/memo/project/util/lifecycle/NotNullMutableLiveData.kt
@@ -1,0 +1,9 @@
+package com.memo.project.util.lifecycle
+
+import androidx.lifecycle.MutableLiveData
+
+class NotNullMutableLiveData<T>(value: T): MutableLiveData<T>(value) {
+    override fun getValue(): T{
+        return super.getValue()!!
+    }
+}
